Move Frame types to frame and frame_options.go
